restapiproj/requirements.txt
Django==4.0
django-rest-framework
pandas==1.3.4